requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
earthengine-api
requests
geopandas
shapely